refactor: 기간 객체 분리
