+++ src/test/java/coupon/coupon/domain/CouponTest.java
@@ -58,21 +58,4 @@ class CouponTest {
                 .isInstanceOf(CouponException.class)
                 .hasMessage("할인율은 3% 이상, 20% 이하이어야 합니다.");
     }
-
-    @DisplayName("종료일이 시작일보다 이전이면 예외가 발생한다.")
-    @Test
-    void cannotCreateIfEndIsBeforeStart() {
-        // given
-        String name = "coupon";
-        int discountAmount = 1000;
-        int minimumOrderAmount = 30000;
-        Category category = Category.FASHION;
-        LocalDate startAt = LocalDate.now();
-        LocalDate endAt = LocalDate.now().minusDays(1);
-
-        // when & then
-        assertThatThrownBy(() -> new Coupon(name, discountAmount, minimumOrderAmount, category, startAt, endAt))
-                .isInstanceOf(CouponException.class)
-                .hasMessage("종료일이 시작일보다 앞설 수 없습니다.");
-    }
 }
